Add files via upload
Source1. using OpenCV to generate cubic element, which will give stepped-like geometry;
Source2(under develop). store OpenCV infor and then identify those elements on the geometry surface. A geometry of inner elements is generated first. after that, those surface elements are later modified and then generated so as to "smooth" the surface;
Source3(to be continued).  An element is generated as soon as one pixel is being detected: if cv:: scalar of the pixel satisfy the threshold, the adjacent pixels (7 other pixels which then become the vertex of mesh element) are examed.